Break cli_common into more specific modules
